API qui march epresque

+++ app/src/main/java/edu/polytech/projet_td2_menu/models/Quantity.java
@@ -2,15 +2,15 @@ package edu.polytech.projet_td2_menu.models;
 
 public class Quantity {
 
-    private final int number;
+    private final double number;
     private final String unite;
 
-    public Quantity(int number, String unite) {
+    public Quantity(double number, String unite) {
         this.number = number;
         this.unite = unite;
     }
 
-    public int getNumber() {
+    public double getNumber() {
         return number;
     }
